+++ src/app/app.component.html
@@ -1,7 +1,4 @@
-<app-welcome-page></app-welcome-page>
-<app-user-login></app-user-login>
-<<<<<<<<< Temporary merge branch 1
-<router-outlet></router-outlet>
-=========
-<app-choose-templet></app-choose-templet>
+
+<!-- <app-user-login></app-user-login> -->
+
 <router-outlet></router-outlet>
