+++ src/app/app.module.ts
@@ -5,13 +5,13 @@ import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { UserLoginComponent } from './userAuth/user-login/user-login.component';
 import { ChooseTempletComponent } from './userAuth/choose-templet/choose-templet.component';
-import { WelcomePageComponent } from './userAuth/welcome-page/welcome-page.component';
+// import { WelcomePageComponent } from './userAuth/welcome-page/welcome-page.component';
 import { DashboardComponent } from './pages/dashboard/dashboard.component';
 import {MatDialogModule} from '@angular/material/dialog'
 import { MatButtonModule } from '@angular/material/button';
 import { MatInputModule } from '@angular/material/input';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-import { PopupDialogComponent } from './userAuth/welcome-page/popup-dialog/popup-dialog.component';
+import { PopupDialogComponent } from './userAuth/choose-templet/popup-dialog/popup-dialog.component';
 import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
 import { MatFormFieldModule } from '@angular/material/form-field';
 
@@ -20,7 +20,7 @@ import { MatFormFieldModule } from '@angular/material/form-field';
     AppComponent,
     UserLoginComponent,
     ChooseTempletComponent,
-    WelcomePageComponent,
+    // WelcomePageComponent,
     DashboardComponent,
     PopupDialogComponent
   ],
