+++ src/app/userAuth/choose-templet/choose-templet.component.ts
@@ -1,5 +1,7 @@
 import { Component } from '@angular/core';
+import { MatDialog } from '@angular/material/dialog';
 import { Router } from '@angular/router';
+import { PopupDialogComponent } from './popup-dialog/popup-dialog.component';
 
 @Component({
   selector: 'app-choose-templet',
@@ -9,7 +11,7 @@ import { Router } from '@angular/router';
 export class ChooseTempletComponent {
 
 
-  constructor(private router:Router){}
+  constructor(private router:Router,private dialog: MatDialog){}
   selectedBox: number | null = null;
   buttonVisible: boolean = false;
 
@@ -18,9 +20,24 @@ export class ChooseTempletComponent {
     this.buttonVisible = true;
   }
 
-  goWelcomePage()
-  {
- this.router.navigate(['welcomePage'])
-    console.log('welcome')
+//   goWelcomePage()
+//   {
+//  this.router.navigate(['welcomePage'])
+//     console.log('welcome')
+//   }
+
+  openDialog(): void {
+    const dialogRef = this.dialog.open(PopupDialogComponent, {
+      width: '1100px',
+      height: '650px',
+      maxWidth:'none',
+      panelClass: 'custom-dialog-container',
+      data: { name: '', email: '' }
+    });
+
+    dialogRef.afterClosed().subscribe(result => {
+      console.log('The dialog was closed');
+      console.log('Form data:', result);
+    });
   }
 }
