--- src/app/userAuth/welcome-page/welcome-page.component.html
@@ -1,2 +1,0 @@
-<!-- app.component.html -->
-<button mat-button (click)="openDialog()">Open Form Dialog</button>
